moved results to separate component from calculator
